changing datatype for refresh func return

+++ src/dds.net-connector-cpp.lib/connector/src/internal/src/variables/var_prim_dword.cpp
@@ -37,7 +37,7 @@ bool
 {
   if (valueProvider != nullptr)
   {
-    int newValue = valueProvider(name);
+    long newValue = valueProvider(name);
 
     if (value != newValue)
     {
